pyproject.toml
[project]
name = "mcp-server-template"
version = "0.1.0"
description = "Python MCP Server template"
readme = "README.md"
requires-python = ">=3.11,<3.14"
authors = [{ name = "Alpic" }]
keywords = ["mcp", "llm", "alpic", "python"]
dependencies = [
    "black>=25.1.0",
    "mcp",
    "python-dotenv>=1.1.0",
    "openai>=1.99.1",
    "instructor==1.0.0",  # Pinned to older version for Docker compatibility
    "requests>=2.31.0",
    "pydantic>=2.11.5",
    # Audio processing packages removed for Docker compatibility
    # Re-enable locally: soundfile, numpy, elevenlabs, kokoro
]
